Create print-reverse-array.cpp
An array is a type of data structure that stores elements of the same type in a contiguous block of memory. In an array,A , of size N, each memory location has some unique index,  (where ),0<=i<N that can be referenced as  (you may also see it written as ).

Given an array, , of  integers, print each element in reverse order as a single line of space-separated integers.