hard
You are given a list(1-indexed) of size , initialized with zeroes. You have to perform  operations on the list and output the maximum of final values of all the  elements in the list. For every operation, you are given three integers ,a,b,k  and  and you have to add value k to all the elements ranging from index a to b(both inclusive).